Implement ability to deep link specific image
Scheme to address individual images in the gallery is implemented with this commit.
The URL is changed accordingly when browsing and the links will work as expected:
they will open the selected image in full viewport mode. The gallery is still
fully loaded in background, though. For this to work, the special URL must be
properly hidden by the webserver, such as using RewriteRule in Apache (see
README for details).